+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Show invalid json error when `NHID` number is empty or less than 14 digits   
 - Explicitly pin OkHttp version to 4.X
 - Use macos version 11.0 to run instrumentation tests on CI
+- Migrate `RecentPatientsScreen` to a fragment
 
 ### Changes
 - Updated translations: `pa-IN`, `hi-IN`, `te-IN`, `kn-IN`, `mr-IN`, `pa-IN`, `te-IN`